fix(authz): second review round - clock-skew-safe passports, the vacuous-every guard, engine split
The sharp three: passports carry a signed iat and the verifier bounds
x against it, so the TTL ceiling holds across nodes with skewed clocks
instead of rejecting valid maximum-TTL tokens; the api-key system-role
guard gains a some() predicate alongside every() - Prisma's every() is
vacuously true on an empty relation, so a bindingless system role
satisfied the exclusivity check for every key on the platform; and the
offboard adapter reads the user's email inside its own transaction, so
an address change between read and delete can't leave a live invite
behind a 'proven complete' offboard (findUserEmail leaves the port).

The rest: engine.ts split into types/scope/matchers/walk siblings, all
under the size limit, with the walk order still literal; the owner
ceiling collects in the same Promise.all as the key's grants (no serial
round trip); bindingNotFound carries only the meta the caller supplied;
the shadow sample-rate parse rejects non-finite input instead of
reading '1oops' as 1; resolveProjectPermissionAny's demo branch fires
its comparison before returning; the boundary test's negative control
is a genuinely independent root; the reader stub and prove callback
are typed against the port so growth fails typecheck; the cache
age-bound test runs on fake timers; the reserved passport secret
leaves .env.example until something reads it; and the thread-share
scenario title now claims exactly what its body asserts.

Suites: authz 54, authz-server 81, app authz/rbac/guards 570; parity
green both branches; biome delta zero-new.
