+++ packages/authz/src/bitset.ts
@@ -3,8 +3,9 @@
  * ordered list, so an effective permission set is a few dozen bytes and a
  * membership test is a bit test. Used by passports; client-safe.
  *
- * The base64url wire codecs live in passport.ts, their only consumer: they
- * need node's Buffer, and this module stays on the browser-safe barrel.
+ * The base64url wire codecs live in passport.ts, which implements the
+ * passport wire format and is their only production consumer: they need
+ * node's Buffer, and this module stays on the browser-safe barrel.
  */
 import { ALL_PERMISSIONS, permissionIndex } from "./registry";
