+++ platform/app/.env.example
@@ -108,7 +108,7 @@ API_TOKEN_JWT_SECRET="change me to a random string"
 
 # AUTHORIZATION
 
-# Rollout knobs for the unified authorization engine (ADR-092). All three are
+# Rollout knobs for the unified authorization engine (ADR-092). Both are
 # OFF when unset, which is the shipped default and the correct setting for
 # every deployment today: the engine runs alongside the existing permission
 # checks, and nothing below changes who can do what. Staged rollout is
@@ -129,13 +129,6 @@ API_TOKEN_JWT_SECRET="change me to a random string"
 # was already taken away. With it off, every check reads from the database,
 # which is always correct.
 # AUTHZ_EPOCH_CACHE=1
-
-# Signing secret for authorization passports — short-lived signed proofs that
-# let a service verify a caller's permissions without a database round trip.
-# Reserved: nothing reads it yet, it ships with the passport rollout. When it
-# does land, generate a dedicated value (never reuse another secret) with:
-#     openssl rand -hex 32
-# AUTHZ_PASSPORT_SECRET=""
 
 # The organization GitHub connection: bot-authored PRs from Langy, and
 # coding-agent pull request linkage. Separate from the GITHUB_CLIENT_*
